Design basic connection APIs for servers (API server and host server).
Design is based around (re)using a single HttpClient for talking to multiple endpoints with gRPC.
The connection manager is built to monitor JUST the connection and not concern itself
with verifying the server identity or even talk gRPC at all.

Servers are identified with endpoints, which are combinations of a uri at which the server
can be reached and a policy which can be used to verify you're reaching the intended server.