Switch to making all fields, except definition, optional on a resource document that comes over the wire.
The definition field is always required so that the resource being operated on can be identified.

The distinction between a fully populated resource document and a (potentially) partial document should
probably be codified in code.
Whenever a resoruce document comes in via the Apply control API it is always partial (it'll always be
missing the state field, or if it's present it'll be ignored).